+++ test/test.js
@@ -2,7 +2,7 @@ const assert = require('assert').strict;
 const { describe, it } = require('mocha');
 const path = require('path');
 const {
-  readCache, writeCache, getAuthors, cacheFor,
+  readCache, writeCache, getAuthors, newCache,
 } = require('..');
 
 const token = process.env.GITHUB_TOKEN;
@@ -19,57 +19,63 @@ const fakeCache = {
   hwy: ['jy.cat@qq.com', 'hwy'],
 };
 
-function p(...filePath) {
-  return path.resolve(__dirname, 'github-file-authors-test', ...filePath);
+function p(...paths) {
+  return path.resolve(__dirname, 'github-file-authors-test', ...paths);
 }
 
 describe('getAuthor', () => {
   it('test/test1 follow without cache', async () => {
-    const authors = await getAuthors({ repo, filePath: p('test', 'test1'), token });
-    assert.deepEqual(authors, ['ouuan', 'sshwy']);
+    const authors = await getAuthors({ repo, paths: p('test', 'test1'), token });
+    assert.deepEqual(authors, [['ouuan', 'sshwy']]);
   });
   it('test/test1 no-follow without cache', async () => {
-    const cache = await cacheFor({ repo, token });
-    assert.deepEqual(cache, new Map());
     const authors = await getAuthors({
-      repo, filePath: p('test', 'test1'), follow: false, cache,
+      repo, paths: [p('test', 'test1')], token, follow: false,
     });
-    assert.deepEqual(authors, ['sshwy']);
+    assert.deepEqual(authors, [['sshwy']]);
   });
   it('test/test1 follow with fake cache', async () => {
     const authors = await getAuthors({
-      repo, filePath: p('test', 'test1'), cache: new Map([fakeCache.nauuo]),
+      repo, paths: [p('test', 'test1')], token, cache: new Map([fakeCache.nauuo]),
     });
-    assert.deepEqual(authors, ['nauuo', 'sshwy']);
+    assert.deepEqual(authors, [['nauuo', 'sshwy']]);
   });
   it('test/test1 no-follow with fake cache', async () => {
     const authors = await getAuthors({
-      repo, filePath: p('test', 'test1'), follow: false, cache: new Map([fakeCache.hwy]),
+      repo, paths: p('test', 'test1'), token, follow: false, cache: new Map([fakeCache.hwy]),
     });
-    assert.deepEqual(authors, ['hwy']);
+    assert.deepEqual(authors, [['hwy']]);
   });
   it('test2/ouuan follow with updated cache', async () => {
-    const cache = await cacheFor({ repo, paths: [p()] });
+    const cache = newCache();
+    await getAuthors({ repo, paths: p(), cache });
     assert.deepEqual(cache, new Map(Object.values(realCache)));
     const authors = await getAuthors({
-      repo, filePath: p('test2', 'ouuan'), cache,
+      repo, paths: p('test2', 'ouuan'), cache,
     });
-    assert.deepEqual(authors, ['Enter-tainer', 'ouuan']);
+    assert.deepEqual(authors, [['Enter-tainer', 'ouuan']]);
   });
   it('test2/ouuan no-follow with updated cache', async () => {
-    const cache = await cacheFor({ repo, paths: [p('test2/ouuan'), p('test/test4')] });
+    const cache = newCache();
+    await getAuthors({ repo, cache, paths: [p('test2/ouuan'), p('test/test4')] });
     assert.deepEqual(cache, new Map([realCache.mgt, realCache.ouuan]));
     const authors = await getAuthors({
-      repo, filePath: p('test2', 'ouuan'), follow: false, cache,
+      repo, paths: [p('test2', 'ouuan')], follow: false, cache,
     });
-    assert.deepEqual(authors, ['Enter-tainer', 'ouuan']);
+    assert.deepEqual(authors, [['Enter-tainer', 'ouuan']]);
+  });
+  it('test/test1 and test2/ouuan', async () => {
+    const authors = await getAuthors({
+      repo, paths: [p('test', 'test1'), p('test2', 'ouuan'), p('test', 'test1'), p('test', 'test1')], token, follow: false,
+    });
+    assert.deepEqual(authors, [['sshwy'], ['Enter-tainer', 'ouuan'], ['sshwy'], ['sshwy']]);
   });
   it('read cache.json', async () => {
     const cache = await readCache(path.resolve(__dirname, 'cache.json'));
     const authors = await getAuthors({
-      repo, filePath: p('test/test2'), token, cache,
+      repo, paths: p('test/test2'), token, cache,
     });
-    assert.deepEqual(authors, ['nauuo', 'sshwy']);
+    assert.deepEqual(authors, [['nauuo', 'sshwy']]);
     assert.deepEqual(cache, new Map([fakeCache.nauuo, realCache.sshwy]));
   });
   it('read/write cache', async () => {
@@ -80,12 +86,12 @@ describe('getAuthor', () => {
     assert.deepEqual(writtenAndReadCache, cache);
   });
   it('Relative path', async () => {
-    const authors = await getAuthors({ repo, filePath: path.relative(process.cwd(), p('test', 'test1')), token });
-    assert.deepEqual(authors, ['ouuan', 'sshwy']);
+    const authors = await getAuthors({ repo, paths: path.relative(process.cwd(), p('test', 'test1')), token });
+    assert.deepEqual(authors, [['ouuan', 'sshwy']]);
   });
   it('On error', async () => {
     let called = false;
-    await cacheFor({
+    await getAuthors({
       repo,
       paths: [p()],
       onerror: (sha, email, error) => {